Refactor OOM support into its own module.
