+++ memapi/src/lib.rs
@@ -15,6 +15,7 @@ use jemallocator::Jemalloc;
 static GLOBAL: Jemalloc = Jemalloc;
 
 mod memorytracking;
+mod oom;
 mod rangemap;
 
 #[no_mangle]
